+++ CHANGELOG.md
@@ -33,6 +33,8 @@
 - VendorDataType
 
 ### Added
+- The Genetics model and it's CRUD endpoints.
+- A global index for the `gene` in the Genetics model.
 - AboutUs page.
 
 ### Changed